+++ NEWS.md
@@ -1,3 +1,3 @@
 # DataMotto 0.1.0
 
-* The first version of DataMotto available at ![https://DataMotto.com](https://DataMotto.com)
+* The first version of DataMotto available at [https://DataMotto.com](https://DataMotto.com)